poprawiony small enemy

+++ Projekt pk4/space invaders ++/Bullet.cpp
@@ -37,6 +37,6 @@ void Bullet::move()
 }
 void Bullet::move(float frametime)
 {
-	position.x +=frametime* (float)speed.x;
-	position.y +=frametime* (float)speed.y;
+	position.x +=(int)(frametime* (float)speed.x);
+	position.y +=(int)(frametime* (float)speed.y);
 }
